# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateLinkBodyType =
  (typeof CreateLinkBodyType)[keyof typeof CreateLinkBodyType];

export const CreateLinkBodyType = {
  send: "send",
  receive: "receive",
} as const;

export type CreateLinkBodyToken =
  (typeof CreateLinkBodyToken)[keyof typeof CreateLinkBodyToken];

export const CreateLinkBodyToken = {
  SOL: "SOL",
  USDC: "USDC",
} as const;

export interface CreateLinkBody {
  type: CreateLinkBodyType;
  /** @nullable */
  recipientAddress?: string | null;
  amountSol: number;
  /** @nullable */
  note?: string | null;
  token: CreateLinkBodyToken;
}

export interface PaymentLink {
  id: string;
  type: string;
  /** @nullable */
  recipientAddress: string | null;
  amountSol: number;
  /** @nullable */
  note: string | null;
  token: string;
  /** @nullable */
  escrowPublicKey: string | null;
  funded: boolean;
  /** @nullable */
  fundedTxSignature: string | null;
  paid: boolean;
  /** @nullable */
  txSignature: string | null;
  /** @nullable */
  payerAddress: string | null;
  /** @nullable */
  claimantAddress: string | null;
  createdAt: string;
  /** @nullable */
  paidAt: string | null;
}

export interface MarkLinkPaidBody {
  txSignature: string;
  payerAddress: string;
}

export interface MarkLinkFundedBody {
  txSignature: string;
}

export interface ClaimLinkBody {
  claimantAddress: string;
}

export interface ClaimLinkResponse {
  txSignature: string;
  claimantAddress: string;
  amountSol: number;
}

export interface ErrorResponse {
  error: string;
}