# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: links
    description: Payment link operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /links:
    post:
      operationId: createLink
      tags: [links]
      summary: Create a payment link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLink"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{linkId}:
    get:
      operationId: getLink
      tags: [links]
      summary: Get a payment link by ID
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLink"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{linkId}/pay:
    patch:
      operationId: markLinkPaid
      tags: [links]
      summary: Mark a receive-type link as paid
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkLinkPaidBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLink"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{linkId}/fund:
    patch:
      operationId: markLinkFunded
      tags: [links]
      summary: Mark a send-type link as funded with escrow tx
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkLinkFundedBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLink"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{linkId}/claim:
    post:
      operationId: claimLink
      tags: [links]
      summary: Claim a send-type link - server sends from escrow to claimant
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimLinkBody"
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimLinkResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateLinkBody:
      type: object
      properties:
        type:
          type: string
          enum: [send, receive]
        recipientAddress:
          type: ["string", "null"]
        amountSol:
          type: number
        note:
          type: ["string", "null"]
        token:
          type: string
          enum: [SOL, USDC]
      required:
        - type
        - amountSol
        - token
    PaymentLink:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        recipientAddress:
          type: ["string", "null"]
        amountSol:
          type: number
        note:
          type: ["string", "null"]
        token:
          type: string
        escrowPublicKey:
          type: ["string", "null"]
        funded:
          type: boolean
        fundedTxSignature:
          type: ["string", "null"]
        paid:
          type: boolean
        txSignature:
          type: ["string", "null"]
        payerAddress:
          type: ["string", "null"]
        claimantAddress:
          type: ["string", "null"]
        createdAt:
          type: string
        paidAt:
          type: ["string", "null"]
      required:
        - id
        - type
        - recipientAddress
        - amountSol
        - note
        - token
        - escrowPublicKey
        - funded
        - fundedTxSignature
        - paid
        - txSignature
        - payerAddress
        - claimantAddress
        - createdAt
        - paidAt
    MarkLinkPaidBody:
      type: object
      properties:
        txSignature:
          type: string
        payerAddress:
          type: string
      required:
        - txSignature
        - payerAddress
    MarkLinkFundedBody:
      type: object
      properties:
        txSignature:
          type: string
      required:
        - txSignature
    ClaimLinkBody:
      type: object
      properties:
        claimantAddress:
          type: string
      required:
        - claimantAddress
    ClaimLinkResponse:
      type: object
      properties:
        txSignature:
          type: string
        claimantAddress:
          type: string
        amountSol:
          type: number
      required:
        - txSignature
        - claimantAddress
        - amountSol
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error