# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a payment link
 */
export const CreateLinkBody = zod.object({
  type: zod.enum(["send", "receive"]),
  recipientAddress: zod.string().nullish(),
  amountSol: zod.number(),
  note: zod.string().nullish(),
  token: zod.enum(["SOL", "USDC"]),
});

/**
 * @summary Get a payment link by ID
 */
export const GetLinkParams = zod.object({
  linkId: zod.coerce.string(),
});

export const GetLinkResponse = zod.object({
  id: zod.string(),
  type: zod.string(),
  recipientAddress: zod.string().nullable(),
  amountSol: zod.number(),
  note: zod.string().nullable(),
  token: zod.string(),
  escrowPublicKey: zod.string().nullable(),
  funded: zod.boolean(),
  fundedTxSignature: zod.string().nullable(),
  paid: zod.boolean(),
  txSignature: zod.string().nullable(),
  payerAddress: zod.string().nullable(),
  claimantAddress: zod.string().nullable(),
  createdAt: zod.string(),
  paidAt: zod.string().nullable(),
});

/**
 * @summary Mark a receive-type link as paid
 */
export const MarkLinkPaidParams = zod.object({
  linkId: zod.coerce.string(),
});

export const MarkLinkPaidBody = zod.object({
  txSignature: zod.string(),
  payerAddress: zod.string(),
});

export const MarkLinkPaidResponse = zod.object({
  id: zod.string(),
  type: zod.string(),
  recipientAddress: zod.string().nullable(),
  amountSol: zod.number(),
  note: zod.string().nullable(),
  token: zod.string(),
  escrowPublicKey: zod.string().nullable(),
  funded: zod.boolean(),
  fundedTxSignature: zod.string().nullable(),
  paid: zod.boolean(),
  txSignature: zod.string().nullable(),
  payerAddress: zod.string().nullable(),
  claimantAddress: zod.string().nullable(),
  createdAt: zod.string(),
  paidAt: zod.string().nullable(),
});

/**
 * @summary Mark a send-type link as funded with escrow tx
 */
export const MarkLinkFundedParams = zod.object({
  linkId: zod.coerce.string(),
});

export const MarkLinkFundedBody = zod.object({
  txSignature: zod.string(),
});

export const MarkLinkFundedResponse = zod.object({
  id: zod.string(),
  type: zod.string(),
  recipientAddress: zod.string().nullable(),
  amountSol: zod.number(),
  note: zod.string().nullable(),
  token: zod.string(),
  escrowPublicKey: zod.string().nullable(),
  funded: zod.boolean(),
  fundedTxSignature: zod.string().nullable(),
  paid: zod.boolean(),
  txSignature: zod.string().nullable(),
  payerAddress: zod.string().nullable(),
  claimantAddress: zod.string().nullable(),
  createdAt: zod.string(),
  paidAt: zod.string().nullable(),
});

/**
 * @summary Claim a send-type link - server sends from escrow to claimant
 */
export const ClaimLinkParams = zod.object({
  linkId: zod.coerce.string(),
});

export const ClaimLinkBody = zod.object({
  claimantAddress: zod.string(),
});

export const ClaimLinkResponse = zod.object({
  txSignature: zod.string(),
  claimantAddress: zod.string(),
  amountSol: zod.number(),
});